+++ gacha/alembic/versions/20260314_0002_banner_cooldowns.py
@@ -13,7 +13,7 @@ depends_on = None
 def upgrade() -> None:
     op.create_table(
         "gacha_player_banner_cooldowns",
-        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column("user_id", sa.BigInteger(), nullable=False),
         sa.Column("banner", sa.String(length=32), nullable=False),
         sa.Column("next_pull_at", sa.DateTime(timezone=True), nullable=True),
         sa.PrimaryKeyConstraint("user_id", "banner"),
